+++ packages/kit/src/runtime/server/page/render.js
@@ -5,6 +5,7 @@ import { serialize_data } from './serialize_data.js';
 import { s } from '../../../utils/misc.js';
 import { Csp } from './csp.js';
 import { stringify_action_response } from './actions.js';
+import { clarify_devalue_error } from '../utils.js';
 
 // TODO rename this function/module
 
@@ -93,7 +94,7 @@ export async function render_response({
 		props.page = {
 			error,
 			params: /** @type {Record<string, any>} */ (event.params),
-			routeId: event.routeId,
+			route: event.route,
 			status,
 			url: event.url,
 			data,
@@ -171,33 +172,33 @@ export async function render_response({
 	const serialized = { data: '', form: 'null' };
 
 	try {
-		serialized.data = devalue.uneval(branch.map(({ server_data }) => server_data));
+		serialized.data = `[${branch
+			.map(({ server_data }) => {
+				if (server_data?.type === 'data') {
+					const data = devalue.uneval(server_data.data);
+
+					const uses = [];
+					if (server_data.uses.dependencies.size > 0) {
+						uses.push(`dependencies:${s(Array.from(server_data.uses.dependencies))}`);
+					}
+
+					if (server_data.uses.params.size > 0) {
+						uses.push(`params:${s(Array.from(server_data.uses.params))}`);
+					}
+
+					if (server_data.uses.parent) uses.push(`parent:1`);
+					if (server_data.uses.route) uses.push(`route:1`);
+					if (server_data.uses.url) uses.push(`url:1`);
+
+					return `{type:"data",data:${data},uses:{${uses.join(',')}}}`;
+				}
+
+				return s(server_data);
+			})
+			.join(',')}]`;
 	} catch (e) {
-		// If we're here, the data could not be serialized with devalue
-		// TODO if we wanted to get super fancy we could track down the origin of the `load`
-		// function, but it would mean passing more stuff around than we currently do
 		const error = /** @type {any} */ (e);
-		const match = /\[(\d+)\]\.data\.(.+)/.exec(error.path);
-		if (match) {
-			throw new Error(
-				`Data returned from \`load\` while rendering ${event.routeId} is not serializable: ${error.message} (data.${match[2]})`
-			);
-		}
-
-		const nonPojoError = /pojo/i.exec(error.message);
-
-		if (nonPojoError) {
-			const constructorName = branch.find(({ server_data }) => server_data?.data?.constructor?.name)
-				?.server_data?.data?.constructor?.name;
-
-			throw new Error(
-				`Data returned from \`load\` (while rendering ${event.routeId}) must be a plain object${
-					constructorName ? ` rather than an instance of ${constructorName}` : ''
-				}`
-			);
-		}
-
-		throw error;
+		throw new Error(clarify_devalue_error(event, error));
 	}
 
 	if (form_value) {
@@ -249,7 +250,7 @@ export async function render_response({
 					error: ${devalue.uneval(error)},
 					node_ids: [${branch.map(({ node }) => node.index).join(', ')}],
 					params: ${devalue.uneval(event.params)},
-					routeId: ${s(event.routeId)},
+					route: ${s(event.route)},
 					data: ${serialized.data},
 					form: ${serialized.form}
 				}` : 'null'},