+++ packages/kit/test/apps/basics/src/routes/before-navigate/prevent-navigation/+page.svelte
@@ -2,12 +2,17 @@
 	import { beforeNavigate } from '$app/navigation';
 
 	let triggered = false;
-	beforeNavigate(({ cancel }) => {
+	let unload = false;
+	let navigation_type;
+	beforeNavigate(({ cancel, type, willUnload }) => {
 		triggered = true;
+		unload = willUnload;
+		navigation_type = type;
 		cancel();
 	});
 </script>
 
 <h1>prevent navigation</h1>
 <a href="/before-navigate/a">a</a>
-<pre>{triggered}</pre>
+<a href="https://google.de">external</a>
+<pre>{triggered} {unload} {navigation_type}</pre>